+++ examples/react-parent-app/src/lib/components/PageTitle.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const PageTitle = styled.h1`
-  font-size: 24px;
-  margin-bottom: 32px;
-  margin-top: 32px;
-`;