+++ packages/common/src/configs.ts
@@ -20,10 +20,9 @@ export const defaultConfigs: ProtocolConfig[] = [
       "https://geth.bsn-development-potassium.bosonportal.io/ac012be65837ebc3134e/rpc",
     theGraphIpfsUrl: "https://ipfs.bsn-development-potassium.bosonportal.io",
     ipfsMetadataUrl: "https://ipfs.bsn-development-potassium.bosonportal.io",
-    widgetsUrl: "https://widgets-test.on.fleek.co",
     contracts: {
-      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/25a707871f4117d29cd5536ef337841509bea2c8
-      protocolDiamond: "0x156207E1ca9746e5a387930c8695d84bc8dAD69F",
+      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/b128b0cc9e6e0a72e6b377573c5ccfa88643e6a5
+      protocolDiamond: "0x559272B7186B3cFe55CE813FB36F52c0E8ad9778",
       testErc20: "0x6aB7d764df1ac4Caa649d3f546e07E2fe44BF667"
     },
     metaTx: {
@@ -40,7 +39,6 @@ export const defaultConfigs: ProtocolConfig[] = [
     jsonRpcUrl: "https://ropsten.infura.io/v3/e8c25128908848db8cb65f595dc0a88f",
     theGraphIpfsUrl: "https://api.thegraph.com/ipfs/api/v0",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
-    widgetsUrl: "https://widgets-staging.on.fleek.co",
     contracts: {
       // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/753ac3aa9cbea4f42236130808984121c65e76dc
       protocolDiamond: "0xDdAE2985Ca872B3E2974e296acF8931C0965801b"
@@ -60,10 +58,9 @@ export const defaultConfigs: ProtocolConfig[] = [
       "https://polygon-mumbai.infura.io/v3/b832a48b9bce4aa6bd9da86eb0126300",
     theGraphIpfsUrl: "https://api.thegraph.com/ipfs/api/v0",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
-    widgetsUrl: "https://widgets-staging.on.fleek.co",
     contracts: {
-      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/dcce13d861b3bee561dc9e06a0e885cb708f7a83
-      protocolDiamond: "0xA2D2D25530900d3BB7855050a164aB80Fed54a89"
+      // from https://github.com/bosonprotocol/boson-protocol-contracts/commit/a24305388c8a2b86cb59667e184677ab8862ebf2
+      protocolDiamond: "0x1796B155D4A719d6eBe0496F4914c98a480e668C"
     },
     metaTx: {
       relayerUrl: "https://api.biconomy.io/api/v2/meta-tx/native",
@@ -78,7 +75,6 @@ export const defaultConfigs: ProtocolConfig[] = [
     subgraphUrl: "",
     jsonRpcUrl: "",
     ipfsMetadataUrl: "https://ipfs.infura.io:5001",
-    widgetsUrl: "", // TODO: replace with prod
     contracts: {
       protocolDiamond: ""
     },
@@ -95,10 +91,9 @@ export const defaultConfigs: ProtocolConfig[] = [
     jsonRpcUrl: "http://127.0.0.1:8545",
     theGraphIpfsUrl: "http://127.0.0.1:5001",
     ipfsMetadataUrl: "http://127.0.0.1:5001",
-    widgetsUrl: "http://127.0.0.1:3000",
     contracts: {
-      protocolDiamond: "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853",
-      testErc20: "0x4c5859f0F772848b2D91F1D83E2Fe57935348029"
+      protocolDiamond: "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6",
+      testErc20: "0xb7278A61aa25c888815aFC32Ad3cC52fF24fE575" // Foreign20 contract
     },
     metaTx: {
       relayerUrl: "http://127.0.0.1:8888",
@@ -107,6 +102,13 @@ export const defaultConfigs: ProtocolConfig[] = [
   }
 ];
 
+/**
+ * Returns default config values based on provided `envName` or `chainId` that can be
+ * used to instantiate a `CoreSDK` instance. The argument `envName` has a higher
+ * specificity than `chainId`.
+ * @param filter - Filter for default config.
+ * @returns Default config.
+ */
 export function getDefaultConfig(filter: {
   envName?: string;
   chainId?: number;