+++ packages/ethers-sdk/src/contracts/factories/ProtocolDiamond__factory.ts
@@ -54,25 +54,6 @@ const _abi = [
     type: "fallback",
   },
   {
-    inputs: [
-      {
-        internalType: "bytes4",
-        name: "_interfaceId",
-        type: "bytes4",
-      },
-    ],
-    name: "supportsInterface",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
     stateMutability: "payable",
     type: "receive",
   },