+++ packages/react-kit/src/index.tsx
@@ -10,5 +10,7 @@ export * from "./components/cta/offer/CommitButton";
 export * from "./components/cta/offer/VoidButton";
 export * from "./components/buttons/Button";
 export * from "./components/searchBar/SearchBar";
+export * from "./components/currencyDisplay/CurrencyDisplay";
+export * from "./components/exchangeCard/ExchangeCard";
 
 export * as hooks from "./hooks";