+++ packages/subgraph/src/mappings/exchange-handler.ts
@@ -5,6 +5,7 @@ import {
   VoucherRevoked,
   VoucherRedeemed,
   VoucherExtended,
+  VoucherTransferred,
   ExchangeCompleted
 } from "../../generated/BosonExchangeHandler/IBosonExchangeHandler";
 import { Exchange, Offer } from "../../generated/schema";
@@ -30,14 +31,14 @@ export function handleBuyerCommittedEvent(event: BuyerCommitted): void {
     saveMetadata(offer, offer.createdAt);
 
     exchange.seller = offer.seller;
-    exchange.validUntilDate = exchangeFromEvent.voucher.validUntilDate;
+    exchange.validUntilDate = event.params.voucher.validUntilDate;
   }
 
   exchange.buyer = exchangeFromEvent.buyerId.toString();
   exchange.offer = exchangeFromEvent.offerId.toString();
   exchange.disputed = false;
   exchange.state = "COMMITTED";
-  exchange.committedDate = event.block.timestamp;
+  exchange.committedDate = event.params.voucher.committedDate;
   exchange.expired = false;
 
   exchange.save();
@@ -86,6 +87,19 @@ export function handleVoucherRedeemedEvent(event: VoucherRedeemed): void {
   if (exchange) {
     exchange.state = "REDEEMED";
     exchange.redeemedDate = event.block.timestamp;
+    exchange.save();
+  }
+}
+
+export function handleVoucherTransferredEvent(event: VoucherTransferred): void {
+  const exchangeId = event.params.exchangeId;
+  const newBuyerId = event.params.newBuyerId;
+
+  const exchange = Exchange.load(exchangeId.toString());
+
+  if (exchange) {
+    exchange.buyer = newBuyerId.toString();
+    // TODO: create transactionLog with all transfers & protocol actions
     exchange.save();
   }
 }